LayoutTests/imported/w3c/web-platform-tests/dom/events/window-event-restored-after-throwing-onerror-expected.txt


PASS window.event is restored after a throwing window.onerror handler
